cli: the CLI now includes their own navigation
The CLI includes their own navigation hierarchy; this allows the
navigation to be in close proximity to the content.  This is especially
useful as it will give fast feedback to CLI documentation writers when
they add a new page, or remove an existing page; the CLI's documentation
handling script(s) can validate the navigation and ensure that it is
up-to-date with the content.

We can then include this navigation and combine it with our own when we
update the content.  This should mean that the documentation builds do
not fail to include the latest checked-in CLI docs.