fix: clarify provenance repository match url is case-sensitive (#1123)
Clarify that the repository url is matched case-sensitive so users don't
need to find out by trial and error.

GitHub user names are case-insensitive which means you can use
`/username/repository` or `/UserName/repository` as valid repository
url. But provenance seems to check for exact match which can be
confusing to people adding provenance to their package on GitHub when
the url does not satisfy provenance.