Fix broken links related to the new "Trusted publishing" page (#1690)
<!-- What / Why -->
<!-- Describe the request in detail. What it does and why it's being
changed. -->

The internal links added in
https://github.com/npm/documentation/pull/1674 followed the on-disk
paths and did not respect that the URL paths use a different scheme, as
explained in
https://github.com/npm/documentation/blob/4c574b89e23515d8411e5ec52624d6a932b874e0/CONTRIBUTING.md#urls:

> ### URLs
> Note that for backward compatibility reasons, the on-disk paths are
not precisely identical to the URLs for the documentation. To keep URLs
expressive but still short, intermediate directories are removed from a
page's URL.
>
> (...)
>
> You might (understandably) expect the page "Creating Node.js modules"
to be URL-wise beneath `contributing-packages-to-the-registry`, but
unfortunately, you would be wrong. To keep URLs short, the intermediate
folder paths are removed from pages, so "Creating Node.js modules"
becomes `https://docs.npmjs.com/creating-node-js-modules`.

Also see
https://github.com/npm/documentation/blob/4c574b89e23515d8411e5ec52624d6a932b874e0/CONTENT-MODEL.md#content-structure:

> ## Content structure
> 
> Docs are grouped by topic.
> 
> - Top-level doc set (Example:
https://docs.npmjs.com/packages-and-modules)
> - Categories (Example:
https://docs.npmjs.com/packages-and-modules/introduction-to-packages-and-modules)
> - **Articles (Example:
https://docs.npmjs.com/about-packages-and-modules)**

As a result, these links did not work at https://docs.npmjs.com/. This
PR fixes them.

## References

Related to https://github.com/npm/documentation/pull/1674 (all links
fixed in this PR were introduced there)

<!-- Examples:
  Related to #0
  Depends on #0
  Blocked by #0
  Fixes #0
  Closes #0
-->