[dec 16th]: Doc Changes for 2FA Publishing and Secure By Default  (#1790)
## Overview
This is a PR to update the documentation to reflect the following
changes coming to the npm registry
1. The default package setting for all newly created packages will be
`Require two-factor authentication or a granular access token with
bypass 2fa enabled`
1. Publishing a package will require the user to have 2fa enabled or use
a granular access token with bypass 2fa enabled.

## Summary of file changes 📖 
- Major updates to **Requiring 2FA for package publishing and settings
modification** to highlight that 2FA or gat is required for publishing.
Package settings options and image is updated.
- Preview:
https://npm-4360decefc-15913497.drafts.github.io/requiring-2fa-for-package-publishing-and-settings-modification
- Adding notice to `npm publish` 
- Preview:
https://npm-4360decefc-15913497.drafts.github.io/cli/v11/commands/npm-publish
- Slight changes to `npm access` to remove option for no 2fa package
setting.
- Preview:
https://npm-4360decefc-15913497.drafts.github.io/cli/v11/commands/npm-access
- Adding a standardized notice to many pages (contributing packages,
organizational settings, about 2 factor authentication)
- Preview:
https://npm-4360decefc-15913497.drafts.github.io/creating-node-js-modules
- Adjusting the wording on **Using private packages in a CI/CD
workflow** to `If trusted publishing is not available for your CI/CD
provider, you must create a [granular access token with bypass 2FA
enabled][create-token] on the website. This will allow you to publish in
your CI/CD workflows even if you have two-factor authentication enabled
on your account.`
- Preview:
https://npm-4360decefc-15913497.drafts.github.io/using-private-packages-in-a-ci-cd-workflow

---------

Co-authored-by: Karen Li <47998177+karenjli@users.noreply.github.com>
Co-authored-by: Nishant Gupta <nishantms@github.com>